move code into view folder
need to seperate some of the tabs into the controller folder
