+++ CMakeLists.txt
@@ -14,10 +14,10 @@ set(
   ${PROJECT_SOURCE_DIR}/threads/command_caller.cc
   ${PROJECT_SOURCE_DIR}/threads/dispatcher_middleman.h
   ${PROJECT_SOURCE_DIR}/tabs/model/status_column_record.cc
-  ${PROJECT_SOURCE_DIR}/tabs/status.cc
-  ${PROJECT_SOURCE_DIR}/tabs/profiles.cc
-  ${PROJECT_SOURCE_DIR}/tabs/processes.cc
-  ${PROJECT_SOURCE_DIR}/tabs/logs.h
+  ${PROJECT_SOURCE_DIR}/tabs/view/status.cc
+  ${PROJECT_SOURCE_DIR}/tabs/view/profiles.cc
+  ${PROJECT_SOURCE_DIR}/tabs/view/processes.cc
+  ${PROJECT_SOURCE_DIR}/tabs/view/logs.h
 )
 
 ### All the code globbed together for easy formatting ###
