+++ src/tabs/view/logs.inl
