+++ src/tabs/view/status.cc
